refactor(review-queue): split REST fallback helpers
Move REST fallback hydration and direct-check helper plumbing out of the oversized review_queue command module so the branch stays under the code-quality LOC ratchet.
